p9_sbe_tracearray -- satsify PRD calls to manage core trace arrays
As a carryover from prior projects, PRD code currently contains logic which
attempts to restart the core trace arrays (via the SBE HWP) after processing
a recoverable error emitted from the core.

The current HWP flags an error in this case (indicating that the core
trace arrays are not SCOM retrievable, which is true for all levels of p9).
This generates a customer visible error log with a FW type callout, which
is undesirable.

This patch is intended to satisfy the current PRD call which intends to reset
and start the core traces, without triggering the check mentioned above or
attempting to access non-implemented SCOM registers.

Ultimately it should have no effect on the actual core tracing, which is
managed on p9 by non-SCOM accessible logic in PC.  I confirmed with Jim Bishop
that the PC logic will not stop tracing on recoverable errors, so there should
be no exposure.

Change-Id: I77e47f71d18b6a3a762ab52b0f6b42d022153f3b
CQ: SW418341
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58078
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58087
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>