PM: Fixed error path handling of getscoms in HWP p9_query_core_access_state.
When is core is in STOP2 or greater, HWP will not be able to SCOM clock
status register of core. Use of FAPI_TRY to scom such a register returns
failure to caller. With thisi, caller fails to understand if output status
variables are good or not. Commit modifies this behavior to always
return SUCCESS and a core state which is consistent with known
status of core.
Key_Cronus_Test=PM_REGRESS

Change-Id: I6365ff0e52c3eb77d9d8ac749f9649a81cdadcac
CQ: SW460324
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76022
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/76035
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>